Add script to deploy to testnet
To run the script, one must first create a `contracts/.env` file with
appropriate values (see `.env.example` for what variables are needed).
Then `script/deploy-for-real.sh` will build the current contracts and
deploy them.

Due to some flakiness on the parts of both Alchemy staging and Sepolia
Arbiscan, this may not exactly work on the first try and may require
fiddling.